Fix a problem with ITC18_SeqExtraParameters
this is used to make NClamp send ITCSeq commands without an extra parameter
that the older drivers required. This was defaulting to 1 which ised the
old form and I could not find a way to make it stick at 0.

This actually fixes what I tried to fix in commit:

d6a9c5c00e50c74bbd0385619be17412a1d0a5a5